Add role to frontend
